requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httptools==0.6.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
typing-extensions==4.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchgod==0.8.2
websockets==11.0.3
